# public/swaggerApi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OTC Forentend API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "{schema}://192.168.10.195:8080",
            "description": "OpenApi parameters",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
        "/otc/google2fa/bind": {
            "get": {
                "tags": [
                    "otc-Google2fa"
                ],
                "summary": "生成秘钥和二维码链接",
                "operationId": "showBindFormGoogle2fa",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "tags": [
                    "otc-Google2fa"
                ],
                "summary": "谷歌二次认证绑定",
                "operationId": "bindGoogle2fa",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "description": "秘钥",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "验证码",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/google2fa/unbind": {
            "post": {
                "tags": [
                    "otc-Google2fa"
                ],
                "summary": "解除谷歌二次验证",
                "operationId": "unbindGoogle2fa",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "验证码",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/country": {
            "get": {
                "tags": [
                    "otc-common"
                ],
                "summary": "国家列表",
                "operationId": "Country",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "data": [
                                            {
                                                "code": "AO",
                                                "areaCode": 244,
                                                "nameEn": "Angola",
                                                "nameZh-CN": "安哥拉",
                                                "nameZh-HK": "安哥拉",
                                                "language": ""
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/currency": {
            "get": {
                "tags": [
                    "otc-common"
                ],
                "summary": "货币列表",
                "operationId": "Currency",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "data": [
                                            {
                                                "code": "AED",
                                                "nameEn": "United Arab Emirates Dirham",
                                                "nameZh-CN": "阿联酋迪拉姆",
                                                "nameZh-HK": "阿聯酋迪拉姆"
                                            },
                                            {
                                                "code": "AFN",
                                                "nameEn": "Afghan Afghani",
                                                "nameZh-CN": "阿富汗尼",
                                                "nameZh-HK": "阿富汗尼"
                                            },
                                            {
                                                "code": "ALL",
                                                "nameEn": "Albania Lek",
                                                "nameZh-CN": "阿尔巴尼列克",
                                                "nameZh-HK": "阿爾巴尼列克"
                                            }
                                        ],
                                        "total": 3
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "otc/config/{configKey}": {
            "get": {
                "tags": [
                    "otc-common"
                ],
                "summary": "系统配置获取",
                "operationId": "otcconfig",
                "parameters": [
                    {
                        "name": "configKey",
                        "in": "query",
                        "description": "OTC配置key:all=全部,counterFee=柜台手续费",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "info": {
                                            "key": "counter_fee",
                                            "type": 1,
                                            "value": "0.7",
                                            "name": "柜台手续费",
                                            "description": "比如手续费 0.7% ，购买 1 BTC 后，实际到账 0.993 BTC 。若是出售，则需要扣除账户 1.007 BTC 的可用余额。",
                                            "rules": "required|numeric|between:0,100",
                                            "weight": 9,
                                            "unit": "%",
                                            "options": null
                                        }
                                    },
                                    "msg": "请求成功"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/otc/blacklist/add/{userId}": {
            "post": {
                "tags": [
                    "otc-ignore"
                ],
                "summary": "添加黑名单",
                "operationId": "addIgnore",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "操作成功",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/blacklist/remove/{userId}": {
            "post": {
                "tags": [
                    "otc-ignore"
                ],
                "summary": "移除黑名单",
                "operationId": "removeIgnore",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "操作成功",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/blacklist/my": {
            "get": {
                "tags": [
                    "otc-ignore"
                ],
                "summary": "我屏蔽的人",
                "operationId": "myIgnore",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "当前页，默认1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "当前页大小，默认10",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "排序字段，默认id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "排列顺序，默认desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "data": [
                                            {
                                                "id": 3,
                                                "fromUid": 950397,
                                                "toUid": 950398,
                                                "tradeCount": 3,
                                                "createTime": "1545974722000",
                                                "updateTime": "1545974722000"
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    }
}